+++ packages/react/src/player/Player.tsx
@@ -27,6 +27,12 @@ interface PlayerProps
   src: Src[] | null;
 
   /**
+   * Sets a custom playback ID for playback.
+   * If not specified, the function defaults to parsing the `src` attribute of the HTMLMediaElement to get the playback ID.
+   */
+  playbackId?: string;
+
+  /**
    * The aspect ratio of the media. Defaults to 16 / 9.
    * This significantly improves cumulative layout shift.
    * Set to `null` to render a plain div primitive.
@@ -53,6 +59,13 @@ interface PlayerProps
    * The interval at which metrics are sent, in ms. Defaults to 5000.
    */
   metricsInterval?: number;
+
+  /**
+   * The ICE servers to use.
+   *
+   * If not provided, the default ICE servers will be used.
+   */
+  iceServers?: RTCIceServer | RTCIceServer[];
 }
 
 const Player = React.memo((props: MediaScopedProps<PlayerProps>) => {
@@ -65,6 +78,7 @@ const Player = React.memo((props: MediaScopedProps<PlayerProps>) => {
     storage,
     onPlaybackEvents,
     metricsInterval,
+    playbackId,
     ...rest
   } = props;
 
@@ -83,6 +97,7 @@ const Player = React.memo((props: MediaScopedProps<PlayerProps>) => {
               },
         ),
       src,
+      playbackId,
       initialProps: {
         aspectRatio,
         jwt,